+++ addons/aax/data/aax_models.hpp
@@ -1,0 +1,4 @@
+class CfgWeapons
+{
+	#include "rolled_sleaves\female_uniform_models.hpp"
+};