+++ addons/aax/data/rolled_sleaves/female_uniform_infos.hpp
@@ -1,0 +1,368 @@
+class Bzn_W_I_CombatUniform
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Urban";
+};
+
+class Bzn_W_I_CombatUniform_Maritime
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Maritime";
+};
+
+class Bzn_W_I_CombatUniform_Woodland
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Woodland";
+};
+
+class Bzn_W_I_CombatUniform_Desert
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Desert";
+};
+
+class Bzn_W_I_CombatUniform_Jungle
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Jungle";
+};
+
+class Bzn_W_I_CombatUniform_Winter
+{
+	model="Bzn_W_I_CombatUniform";
+	element="King";
+	camo="Winter";
+};
+
+class Bzn_W_I_CombatUniform_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Urban";
+};
+
+class Bzn_W_I_CombatUniform_Maritime_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Maritime";
+};
+
+class Bzn_W_I_CombatUniform_Woodland_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Woodland";
+};
+
+class Bzn_W_I_CombatUniform_Desert_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Desert";
+};
+
+class Bzn_W_I_CombatUniform_Jungle_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Jungle";
+};
+
+class Bzn_W_I_CombatUniform_Winter_Knight
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Knight";
+	camo="Winter";
+};
+
+class Bzn_W_I_CombatUniform_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Urban";
+};
+
+class Bzn_W_I_CombatUniform_Maritime_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Maritime";
+};
+
+class Bzn_W_I_CombatUniform_Woodland_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Woodland";
+};
+
+class Bzn_W_I_CombatUniform_Desert_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Desert";
+};
+
+class Bzn_W_I_CombatUniform_Jungle_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Jungle";
+};
+
+class Bzn_W_I_CombatUniform_Winter_Queen
+{
+	model="Bzn_W_I_CombatUniform";
+	element="Queen";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_RS_F_Urban
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_RS_F_Winter
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_RS_F_Desert
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Desert";
+};
+
+class Bzn_Field_Uniform_RS_F_Jungle
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Jungle";
+};
+
+class Bzn_Field_Uniform_RS_F_Woodland
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Woodland";
+};
+
+class Bzn_Field_Uniform_RS_F_Maritime
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="King";
+	camo="Maritime";
+};
+
+class Bzn_Field_Uniform_RS_F_Urban_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_RS_F_Winter_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_RS_F_Desert_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Desert";
+};
+
+class Bzn_Field_Uniform_RS_F_Jungle_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Jungle";
+};
+
+class Bzn_Field_Uniform_RS_F_Woodland_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Woodland";
+};
+
+class Bzn_Field_Uniform_RS_F_Maritime_Queen
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Queen";
+	camo="Maritime";
+};
+
+class Bzn_Field_Uniform_RS_F_Urban_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_RS_F_Winter_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_RS_F_Desert_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Desert";
+};
+
+class Bzn_Field_Uniform_RS_F_Jungle_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Jungle";
+};
+
+class Bzn_Field_Uniform_RS_F_Woodland_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Woodland";
+};
+
+class Bzn_Field_Uniform_RS_F_Maritime_Knight
+{
+	model="Bzn_Field_Uniform_RS_F_Urban";
+	element="Knight";
+	camo="Maritime";
+};
+
+class Bzn_Field_Uniform_F_Urban
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_F_Winter
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_F_Desert
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Desert";
+};
+class Bzn_Field_Uniform_F_Jungle
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Jungle";
+};
+class Bzn_Field_Uniform_F_Woodland
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Woodland";
+};
+class Bzn_Field_Uniform_F_Maritime
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="King";
+	camo="Maritime";
+};
+
+class Bzn_Field_Uniform_F_Urban_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_F_Winter_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_F_Desert_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Desert";
+};
+class Bzn_Field_Uniform_F_Jungle_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Jungle";
+};
+class Bzn_Field_Uniform_F_Woodland_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Woodland";
+};
+class Bzn_Field_Uniform_F_Maritime_Queen
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Queen";
+	camo="Maritime";
+};
+
+class Bzn_Field_Uniform_F_Urban_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Urban";
+};
+
+class Bzn_Field_Uniform_F_Winter_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Winter";
+};
+
+class Bzn_Field_Uniform_F_Desert_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Desert";
+};
+class Bzn_Field_Uniform_F_Jungle_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Jungle";
+};
+class Bzn_Field_Uniform_F_Woodland_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Woodland";
+};
+class Bzn_Field_Uniform_F_Maritime_Knight
+{
+	model="Bzn_Field_Uniform_F_Urban";
+	element="Knight";
+	camo="Maritime";
+};