+++ addons/aax/data/rolled_sleaves/female_uniform_models.hpp
@@ -1,0 +1,98 @@
+class Bzn_W_I_CombatUniform
+{
+	label="[Bzn] Female Uniforms";
+	author="Bzn Mod Dev Team, Yandere";
+	options[]={"element","camo"};
+	class element
+	{
+		label="Element";
+		alwaysSelectable=1;
+		changeingame = 0;
+		values[]={"King","Queen","Knight"};
+		class King
+		{
+			image="z\bzn\addons\aax\data\images\King.jpg";
+		};
+		class Queen
+		{
+			image="z\bzn\addons\aax\data\images\Queen.jpg";
+		};
+		class Knight
+		{
+			image="z\bzn\addons\aax\data\images\Knight.jpg";
+		};
+	};
+	class camo
+	{
+		label="Camo";
+		alwaysSelectable=1;
+		changeingame=0;
+		values[]={"Urban","Desert","Winter","Woodland","Jungle","Maritime"};
+	};
+};
+
+class Bzn_Field_Uniform_RS_F_Urban
+{
+	label="[Bzn] Uniforms (Rolled)";
+	author="Bzn Mod Dev Team, Yandere";
+	options[]={"element","camo"};
+	class element
+	{
+		label="Element";
+		alwaysSelectable=1;
+		changeingame = 0;
+		values[]={"King","Queen","Knight"};
+		class King
+		{
+			image="z\bzn\addons\aax\data\images\King.jpg";
+		};
+		class Queen
+		{
+			image="z\bzn\addons\aax\data\images\Queen.jpg";
+		};
+		class Knight
+		{
+			image="z\bzn\addons\aax\data\images\Knight.jpg";
+		};
+	};
+	class camo
+	{
+		label="Camo";
+		alwaysSelectable=1;
+		changeingame=0;
+		values[]={"Urban","Desert","Winter","Woodland","Jungle","Maritime"};
+	};
+};
+
+class Bzn_Field_Uniform_F_Urban
+{
+	label="[Bzn] Uniforms";
+	author="Bzn Mod Dev Team, Yandere";
+	options[]={"element","camo"};
+	class element
+	{
+		label="Element";
+		alwaysSelectable=1;
+		changeingame = 0;
+		values[]={"King","Queen","Knight"};
+		class King
+		{
+			image="z\bzn\addons\aax\data\images\King.jpg";
+		};
+		class Queen
+		{
+			image="z\bzn\addons\aax\data\images\Queen.jpg";
+		};
+		class Knight
+		{
+			image="z\bzn\addons\aax\data\images\Knight.jpg";
+		};
+	};
+	class camo
+	{
+		label="Camo";
+		alwaysSelectable=1;
+		changeingame=0;
+		values[]={"Urban","Desert","Winter","Woodland","Jungle","Maritime"};
+	};
+};